Move fake message handle to test code
